project/learning_openai_connect/backend/pyproject.toml
[project]
name = "learning-openai-connect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "janome>=0.5.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "openai>=1.85.0",
    "openai-whisper",
    "pykakasi>=2.3.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "rapidfuzz>=3.13.0",
    "scipy>=1.15.3",
    "sounddevice>=0.5.2",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.3",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
